Time: 20 ms (11.21%) | Memory: 17.3 MB (44.86%) - LeetSync

+++ 788-minimize-max-distance-to-gas-station/minimize-max-distance-to-gas-station.cpp
@@ -116,6 +116,6 @@ public:
             }
         }
 
-        return (double)high;
+        return high;
     }
 };
